pyproject.toml
[project]
name = "teacher-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "chainlit>=2.4.301",
    "instructor>=1.7.9",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.2",
    "pydantic-ai>=0.0.52",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.4",
]

[tool.ruff]
# Basic settings
line-length = 88
target-version = "py312"

# Essential rules for your current codebase
[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
]

# Directories to ignore
exclude = [
    ".git",
    ".venv",
    "__pycache__",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
markers = [
    "llm: marks tests as LLM-based (deselect with '-m \"not llm\"')",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
